requirements.txt
streamlit
pandas
plotly
loguru
crewai # For agent orchestration (even if simulated via JSON for demo)
python-dotenv # For loading .env
openai  # For Azure OpenAI compatibility